broke out the onboarding modal into its own view
